# everis.ovsdb-rest/target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"EVERIS-OVSDB","description":"EVERIS-OVSDB API to make control over the switches using OVSDB Protocol"},"basePath":"/onos/everis.ovsdb-rest","produces":["application/json"],"consumes":["application/json"],"tags":[{"name":"config","description":"Possible REST APIs to make changes in the switches of the ONOS Cluster"}],"paths":{"/config/sample":{"get":{"tags":["config"],"summary":"Get hello world greeting","description":"","parameters":[],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/config/createBridge":{"post":{"tags":["config"],"summary":"Create a Bridge using JSON","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/create_delete_Bridge"},"description":"JSON Parameter","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/config/deletePort":{"post":{"tags":["config"],"summary":"Delete the port of a bridge","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/add_delete_Port"},"description":"JSON Parameter","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/config/createPatchPeerPort":{"post":{"tags":["config"],"summary":"Creates Patch Peer Port to connect with another port","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/createPatchPeerPort"},"description":"JSON Configuration","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/config/createPathIntent":{"post":{"tags":["config"],"summary":"Create a Path Intent for a known Path Port","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/PathIntent"},"description":"JSON Configuration to make the PathIntent","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/config/deleteBridge":{"post":{"tags":["config"],"summary":"Delete a Bridge using JSON","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/create_delete_Bridge"},"description":"JSON Parameter","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/config/addPort":{"post":{"tags":["config"],"summary":"Add a port in the bridge with the given information","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/add_delete_Port"},"description":"JSON Parameter","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/config/createVxlanTunnel":{"post":{"tags":["config"],"summary":"Create a VXLAN Tunnel Port","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/addVxlanTunnel"},"description":"JSON Configuration","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}}},"definitions":{"create_delete_Bridge":{"type":"object","required":["ovsdb-ip","bridge-name"],"properties":{"ovsdb-ip":{"type":"string","format":"IpAddress","example":"35.232.247.116"},"bridge-name":{"type":"string","format":"srt","example":"br-1"}}},"add_delete_Port":{"type":"object","required":["ovsdb-ip","bridge-name","port-name"],"properties":{"ovsdb-ip":{"type":"string","format":"IpAddress","example":"35.232.247.116"},"bridge-name":{"type":"string","format":"str","example":"br-1"},"port-name":{"type":"string","format":"srt","example":"port1"}}},"createPatchPeerPort":{"type":"object","required":["ovsdb-ip","bridge-name","port-name","patch-peer"],"properties":{"ovsdb-ip":{"type":"string","format":"IpAddress","example":"35.232.247.116"},"bridge-name":{"type":"string","format":"str","example":"br-1"},"port-name":{"type":"string","format":"srt","example":"port1"},"patch-peer":{"type":"string","format":"srt","example":"patchPort"}}},"addVxlanTunnel":{"type":"object","required":["ovsdb-ip","bridge-name","port-name","remote-ip","key"],"properties":{"ovsdb-ip":{"type":"string","format":"IpAddress","example":"35.232.247.116"},"bridge-name":{"type":"string","format":"str","example":"br-1"},"port-name":{"type":"string","format":"srt","example":"port1"},"remote-ip":{"type":"string","format":"srt","example":"192.168.0.100"},"key":{"type":"string","format":"srt","example":"flow"}}},"PathIntent":{"type":"object","required":["src-id","dst-id","port-src","port-dst","setType"],"properties":{"src-id":{"type":"string","format":"str","example":"of:0000005000000202"},"dst-id":{"type":"string","format":"str","example":"of:0000005000000102"},"port-src":{"type":"string","format":"srt","example":"3"},"port-dst":{"type":"string","format":"srt","example":"3"},"setType":{"type":"string","format":"srt","example":"PRIMARY/BACKUP/FAILOVER"}}}}}